Small refactor of the test results component.
