+++ frontend/src/app/testResults/TestResultsList.test.tsx
@@ -15,18 +15,20 @@ import userEvent from "@testing-library/user-event";
 
 import {
   GetAllFacilitiesDocument,
-  GetFacilityResultsForCsvDocument,
   GetFacilityResultsMultiplexDocument,
   GetResultsCountByFacilityDocument,
 } from "../../generated/graphql";
-import { QUERY_PATIENT } from "../testQueue/addToQueue/AddToQueueSearch";
 import { appPermissions } from "../permissions";
 
-import { testResultDetailsQuery } from "./TestResultDetailsModal";
 import TestResultsList, {
   ALL_FACILITIES_ID,
   DetachedTestResultsList,
 } from "./TestResultsList";
+import testResultsMultiplex from "./mocks/resultsMultiplex";
+import testResults from "./mocks/resultsCovid";
+import testResultsByStartDateAndEndDate from "./mocks/resultsByStartAndEndDate";
+import mocks from "./mocks/queries";
+import { facilities, facilitiesIncludeArchived } from "./mocks/facilities";
 
 const mockStore = configureStore([]);
 const store = mockStore({
@@ -58,1275 +60,6 @@ const WithRouter: React.FC = ({ children }) => (
     {children}
   </MemoryRouter>
 );
-
-// Data copied from Chrome network window
-const testResults = [
-  {
-    internalId: "0969da96-b211-41cd-ba61-002181f0918d",
-    dateTested: "2021-03-17T19:27:23.806Z",
-    result: "NEGATIVE",
-    results: [{ disease: { name: "COVID-19" }, testResult: "NEGATIVE" }],
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
-      firstName: "Barb",
-      middleName: "Whitaker",
-      lastName: "Cragell",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Arthur",
-        middleName: "A",
-        lastName: "Admin",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: true,
-    symptoms: "{}",
-    __typename: "TestResult",
-  },
-  {
-    internalId: "7c768a5d-ef90-44cd-8050-b96dd77f51d5",
-    dateTested: "2021-03-18T19:27:21.052Z",
-    result: "NEGATIVE",
-    results: [{ disease: { name: "COVID-19" }, testResult: "NEGATIVE" }],
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "f74ad245-3a69-44b5-bb6d-efe06308bb85",
-      firstName: "Barde",
-      middleName: "X",
-      lastName: "Colleer",
-      birthDate: "1960-11-07",
-      gender: "female",
-      role: "STAFF",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Ursula",
-        middleName: "",
-        lastName: "User",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: false,
-    symptoms: "{}",
-    __typename: "TestResult",
-  },
-  {
-    internalId: "7c768a5d-ef90-44cd-8050-b96dd7aaa1d5",
-    dateTested: "2021-03-19T19:27:21.052Z",
-    result: "POSITIVE",
-    results: [{ disease: { name: "COVID-19" }, testResult: "POSITIVE" }],
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "f74ad245-3a69-44b5-bb6d-efe06308bb85",
-      firstName: "Sam",
-      middleName: "G",
-      lastName: "Gerard",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      lookupId: null,
-      email: "sam@gerard.com",
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Ethan",
-        middleName: "",
-        lastName: "Entry",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: false,
-    symptoms: '{"someSymptom":"true"}',
-    __typename: "TestResult",
-  },
-];
-
-const testResultsMultiplex = [
-  {
-    internalId: "0969da96-b211-41cd-ba61-002181f0918d",
-    dateTested: "2021-03-17T19:27:23.806Z",
-    result: "NEGATIVE",
-    results: [
-      { disease: { name: "COVID-19" }, testResult: "NEGATIVE" },
-      { disease: { name: "Flu A" }, testResult: "POSITIVE" },
-      { disease: { name: "Flu B" }, testResult: "NEGATIVE" },
-    ],
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "48c523e8-7c65-4047-955c-e3f65bb8b58b",
-      firstName: "Gruff",
-      middleName: "M",
-      lastName: "MacGuffin",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Arthur",
-        middleName: "A",
-        lastName: "Admin",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: true,
-    symptoms: "{}",
-    __typename: "TestResult",
-  },
-  {
-    internalId: "7c768a5d-ef90-44cd-8050-b96dd77f51d5",
-    dateTested: "2021-03-18T19:27:21.052Z",
-    result: "NEGATIVE",
-    results: [
-      { disease: { name: "COVID-19" }, testResult: "NEGATIVE" },
-      { disease: { name: "Flu A" }, testResult: "NEGATIVE" },
-      { disease: { name: "Flu B" }, testResult: "POSITIVE" },
-    ],
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "f74ad245-3a69-44b5-bb6d-efe06308bb85",
-      firstName: "Braisley",
-      middleName: "R",
-      lastName: "Adams",
-      birthDate: "1960-11-07",
-      gender: "female",
-      role: "STAFF",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Ursula",
-        middleName: "",
-        lastName: "User",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: false,
-    symptoms: "{}",
-    __typename: "TestResult",
-  },
-  {
-    internalId: "7c768a5d-ef90-44cd-8050-b96dd7aaa1d5",
-    dateTested: "2021-03-19T19:27:21.052Z",
-    result: "POSITIVE",
-    results: [],
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "f74ad245-3a69-44b5-bb6d-efe06308bb85",
-      firstName: "Rupert",
-      middleName: "G",
-      lastName: "Purrington",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      lookupId: null,
-      email: "sam@gerard.com",
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Ethan",
-        middleName: "",
-        lastName: "Entry",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: false,
-    symptoms: '{"someSymptom":"true"}',
-    __typename: "TestResult",
-  },
-];
-
-const testResultsForCsv = [
-  {
-    facility: { name: "Central Middle School", __typename: "Facility" },
-    dateTested: "2022-01-19T16:45:11.446Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    reasonForCorrection: null,
-    deviceType: {
-      name: "Cue",
-      manufacturer: "CueTM",
-      model: "Cue COVID Test",
-      swabType: "45345873",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
-      firstName: "Barb",
-      middleName: "Whitaker",
-      lastName: "Cragell",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      race: "asian",
-      ethnicity: "not_hispanic",
-      tribalAffiliation: [null],
-      lookupId: null,
-      telephone: "(905) 594-3725",
-      email: "fimevusuwu@mailinator.com",
-      street: "67 White Oak Avenue",
-      streetTwo: "Voluptatem optio in",
-      city: "Magna ut corrupti s",
-      county: "Maiores et sunt et q",
-      state: "NA",
-      zipCode: "00000",
-      country: null,
-      residentCongregateSetting: null,
-      employedInHealthcare: null,
-      preferredLanguage: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Nathan",
-        middleName: null,
-        lastName: "Carter",
-        __typename: "NameInfo",
-      },
-      __typename: "ApiUser",
-    },
-    symptoms:
-      '{"64531003":"false","103001002":"false","84229001":"false","68235000":"false","426000000":"false","49727002":"false","68962001":"false","422587007":"false","267036007":"false","62315008":"false","43724002":"false","36955009":"false","44169009":"false","422400008":"false","230145002":"false","25064002":"false","162397003":"false"}',
-    noSymptoms: false,
-    symptomOnset: null,
-    __typename: "TestResult",
-  },
-  {
-    facility: { name: "Central Middle School", __typename: "Facility" },
-    dateTested: "2022-01-19T16:42:46.744Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    reasonForCorrection: null,
-    deviceType: {
-      name: "Cue",
-      manufacturer: "CueTM",
-      model: "Cue COVID Test",
-      swabType: "45345873",
-      __typename: "DeviceType",
-    },
-    patient: {
-      firstName: "Cadman",
-      middleName: "Burton Conner",
-      lastName: "Day",
-      birthDate: "1903-04-07",
-      gender: "male",
-      race: "asian",
-      ethnicity: "not_hispanic",
-      tribalAffiliation: [null],
-      lookupId: null,
-      telephone: "(905) 594-3725",
-      email: "fimevusuwu@mailinator.com",
-      street: "67 White Oak Avenue",
-      streetTwo: "Voluptatem optio in",
-      city: "Magna ut corrupti s",
-      county: "Maiores et sunt et q",
-      state: "NA",
-      zipCode: "00000",
-      country: null,
-      role: "STUDENT",
-      residentCongregateSetting: null,
-      employedInHealthcare: null,
-      preferredLanguage: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Nathan",
-        middleName: null,
-        lastName: "Carter",
-        __typename: "NameInfo",
-      },
-      __typename: "ApiUser",
-    },
-    symptoms:
-      '{"64531003":"false","103001002":"false","84229001":"false","68235000":"false","426000000":"false","49727002":"false","68962001":"false","422587007":"false","267036007":"false","62315008":"false","43724002":"false","36955009":"false","44169009":"false","422400008":"false","230145002":"false","25064002":"false","162397003":"false"}',
-    noSymptoms: false,
-    symptomOnset: null,
-    __typename: "TestResult",
-  },
-  {
-    facility: { name: "Central Middle School", __typename: "Facility" },
-    dateTested: "2022-01-13T22:44:52.193Z",
-    result: "POSITIVE",
-    correctionStatus: "ORIGINAL",
-    reasonForCorrection: null,
-    deviceType: {
-      name: "Cue",
-      manufacturer: "CueTM",
-      model: "Cue COVID Test",
-      swabType: "45345873",
-      __typename: "DeviceType",
-    },
-    patient: {
-      firstName: "Bob",
-      middleName: null,
-      lastName: "International",
-      birthDate: "1970-09-09",
-      gender: "male",
-      race: null,
-      ethnicity: null,
-      tribalAffiliation: [null],
-      lookupId: null,
-      telephone: "(530) 867-5309",
-      email: "zedd@skylight.digital",
-      street: "123 Main St",
-      streetTwo: "",
-      city: null,
-      county: null,
-      state: "CA",
-      zipCode: "95820",
-      country: "USA",
-      role: "UNKNOWN",
-      residentCongregateSetting: null,
-      employedInHealthcare: null,
-      preferredLanguage: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Zedd",
-        middleName: null,
-        lastName: "Shmais",
-        __typename: "NameInfo",
-      },
-      __typename: "ApiUser",
-    },
-    symptoms:
-      '{"64531003":"false","103001002":"false","84229001":"false","68235000":"false","426000000":"false","49727002":"false","68962001":"false","422587007":"false","267036007":"false","62315008":"false","43724002":"false","36955009":"false","44169009":"false","422400008":"false","230145002":"false","25064002":"false","162397003":"false"}',
-    noSymptoms: false,
-    symptomOnset: null,
-    __typename: "TestResult",
-  },
-];
-
-const testResultsByPatient = [
-  {
-    internalId: "0969da96-b211-41cd-ba61-002181f0918d",
-    dateTested: "2021-03-17T19:27:23.806Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
-      firstName: "Barb",
-      middleName: "Whitaker",
-      lastName: "Cragell",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Arthur",
-        middleName: "A",
-        lastName: "Admin",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: true,
-    symptoms: "{}",
-    __typename: "TestResult",
-  },
-];
-
-const testResultsByResultValue = [
-  {
-    internalId: "0969da96-b211-41cd-ba61-002181f0918d",
-    dateTested: "2021-03-17T19:27:23.806Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
-      firstName: "Barb",
-      middleName: "Whitaker",
-      lastName: "Cragell",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Arthur",
-        middleName: "A",
-        lastName: "Admin",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: true,
-    symptoms: "{}",
-    __typename: "TestResult",
-  },
-  {
-    internalId: "7c768a5d-ef90-44cd-8050-b96dd77f51d5",
-    dateTested: "2021-03-18T19:27:21.052Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "f74ad245-3a69-44b5-bb6d-efe06308bb85",
-      firstName: "Barde",
-      middleName: "X",
-      lastName: "Colleer",
-      birthDate: "1960-11-07",
-      gender: "female",
-      role: "STAFF",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Ursula",
-        middleName: "",
-        lastName: "User",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: false,
-    symptoms: "{}",
-    __typename: "TestResult",
-  },
-];
-
-const testResultsByRole = [
-  {
-    internalId: "0969da96-b211-41cd-ba61-002181f0918d",
-    dateTested: "2021-03-17T19:27:23.806Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
-      firstName: "Barb",
-      middleName: "Whitaker",
-      lastName: "Cragell",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Arthur",
-        middleName: "A",
-        lastName: "Admin",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: true,
-    symptoms: "{}",
-    __typename: "TestResult",
-  },
-  {
-    internalId: "7c768a5d-ef90-44cd-8050-b96dd7aaa1d5",
-    dateTested: "2021-03-19T19:27:21.052Z",
-    result: "POSITIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "f74ad245-3a69-44b5-bb6d-efe06308bb85",
-      firstName: "Sam",
-      middleName: "G",
-      lastName: "Gerard",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Ethan",
-        middleName: "",
-        lastName: "Entry",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: false,
-    symptoms: '{"someSymptom":"true"}',
-    __typename: "TestResult",
-  },
-];
-
-const testResultsByStartDate = [
-  {
-    internalId: "7c768a5d-ef90-44cd-8050-b96dd77f51d5",
-    dateTested: "2021-03-18T19:27:21.052Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "f74ad245-3a69-44b5-bb6d-efe06308bb85",
-      firstName: "Barde",
-      middleName: "X",
-      lastName: "Colleer",
-      birthDate: "1960-11-07",
-      gender: "female",
-      role: "STAFF",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Ursula",
-        middleName: "",
-        lastName: "User",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: false,
-    symptoms: "{}",
-    __typename: "TestResult",
-  },
-  {
-    internalId: "7c768a5d-ef90-44cd-8050-b96dd7aaa1d5",
-    dateTested: "2021-03-19T19:27:21.052Z",
-    result: "POSITIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "f74ad245-3a69-44b5-bb6d-efe06308bb85",
-      firstName: "Sam",
-      middleName: "G",
-      lastName: "Gerard",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Ethan",
-        middleName: "",
-        lastName: "Entry",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: false,
-    symptoms: '{"someSymptom":"true"}',
-    __typename: "TestResult",
-  },
-];
-
-const testResultsByStartDateAndEndDate = [
-  {
-    internalId: "7c768a5d-ef90-44cd-8050-b96dd77f51d5",
-    dateTested: "2021-03-18T19:27:21.052Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "f74ad245-3a69-44b5-bb6d-efe06308bb85",
-      firstName: "Barde",
-      middleName: "X",
-      lastName: "Colleer",
-      birthDate: "1960-11-07",
-      gender: "female",
-      role: "STAFF",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Ursula",
-        middleName: "",
-        lastName: "User",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    noSymptoms: false,
-    symptoms: "{}",
-    __typename: "TestResult",
-  },
-];
-
-const testResultsByFacility = [
-  {
-    internalId: "0969da96-b211-41cd-ba61-002181f0123a",
-    dateTested: "2021-04-12T12:40:33.381Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "48c523e8-7c65-4047-955c-e3f65bb8123a",
-      firstName: "Lewis",
-      middleName: "",
-      lastName: "Clarkson",
-      birthDate: "1958-08-25",
-      gender: "other",
-      role: "VISITOR",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Arthur",
-        middleName: "A",
-        lastName: "Admin",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8123a",
-    },
-    facility: {
-      name: "Facility 2",
-    },
-    __typename: "TestResult",
-  },
-];
-
-const testResultsByAllFacility = [
-  {
-    internalId: "0969da96-b211-41cd-ba61-002181f0123a",
-    dateTested: "2021-04-12T12:40:33.381Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "48c523e8-7c65-4047-955c-e3f65bb8123a",
-      firstName: "Lewis",
-      middleName: "",
-      lastName: "Clarkson",
-      birthDate: "1958-08-25",
-      gender: "other",
-      role: "VISITOR",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Arthur",
-        middleName: "A",
-        lastName: "Admin",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8123a",
-    },
-    facility: {
-      name: "Facility 2",
-    },
-    __typename: "TestResult",
-  },
-  {
-    internalId: "0969da96-b211-41cd-ba61-002181f0918d",
-    dateTested: "2021-03-17T19:27:23.806Z",
-    result: "NEGATIVE",
-    correctionStatus: "ORIGINAL",
-    deviceType: {
-      internalId: "8c1a8efe-8951-4f84-a4c9-dcea561d7fbb",
-      name: "Abbott IDNow",
-      __typename: "DeviceType",
-    },
-    patient: {
-      internalId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
-      firstName: "Barb",
-      middleName: "Whitaker",
-      lastName: "Cragell",
-      birthDate: "1960-11-07",
-      gender: "male",
-      role: "RESIDENT",
-      lookupId: null,
-      __typename: "Patient",
-    },
-    createdBy: {
-      nameInfo: {
-        firstName: "Arthur",
-        middleName: "A",
-        lastName: "Admin",
-      },
-    },
-    patientLink: {
-      internalId: "68c543e8-7c65-4047-955c-e3f65bb8b58a",
-    },
-    facility: {
-      name: "Facility 1",
-    },
-    __typename: "TestResult",
-  },
-];
-
-const patients = [
-  {
-    internalId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
-    firstName: "Barb",
-    middleName: "Whitaker",
-    lastName: "Cragell",
-    birthDate: "1960-11-07",
-    gender: "male",
-    role: "RESIDENT",
-    lookupId: null,
-    __typename: "Patient",
-  },
-];
-
-const facilities = [
-  {
-    id: "1",
-    name: "Facility 1",
-    isDeleted: false,
-  },
-  {
-    id: "2",
-    name: "Facility 2",
-    isDeleted: false,
-  },
-];
-
-const facilitiesIncludeArchived = facilities.concat({
-  id: "3",
-  name: "Facility 3",
-  isDeleted: true,
-});
-
-const mocks = [
-  {
-    request: {
-      query: GetResultsCountByFacilityDocument,
-      variables: {
-        facilityId: "1",
-      },
-    },
-    result: {
-      data: {
-        testResultsCount: testResults.length,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsMultiplexDocument,
-      variables: {
-        facilityId: "1",
-        pageNumber: 0,
-        pageSize: 20,
-      },
-    },
-    result: {
-      data: {
-        testResults,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsForCsvDocument,
-      variables: {
-        facilityId: "1",
-        pageNumber: 0,
-        pageSize: 3,
-      },
-    },
-    result: {
-      data: {
-        testResults: testResultsForCsv,
-      },
-    },
-  },
-  {
-    request: {
-      query: testResultDetailsQuery,
-      variables: {
-        id: testResults[0].internalId,
-      },
-    },
-    result: {
-      data: {
-        testResult: {
-          dateTested: "2021-03-17T19:27:23.806Z",
-          result: "NEGATIVE",
-          correctionStatus: "ORIGINAL",
-          deviceType: {
-            name: "Abbott IDNow",
-            __typename: "DeviceType",
-          },
-          patient: {
-            firstName: "Barb",
-            middleName: "Whitaker",
-            lastName: "Cragell",
-            birthDate: "1960-11-07",
-          },
-          createdBy: {
-            name: {
-              firstName: "Arthur",
-              middleName: "A",
-              lastName: "Admin",
-            },
-          },
-          symptoms: "{}",
-          symptomOnset: null,
-          pregnancy: null,
-          __typename: "TestResult",
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: GetResultsCountByFacilityDocument,
-      variables: {
-        facilityId: "1",
-        patientId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
-      },
-    },
-    result: {
-      data: {
-        testResultsCount: testResultsByPatient.length,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsMultiplexDocument,
-      variables: {
-        facilityId: "1",
-        patientId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
-        pageNumber: 0,
-        pageSize: 20,
-      },
-    },
-    result: {
-      data: {
-        testResults: testResultsByPatient,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetResultsCountByFacilityDocument,
-      variables: {
-        facilityId: "1",
-        result: "NEGATIVE",
-      },
-    },
-    result: {
-      data: {
-        testResultsCount: testResultsByResultValue.length,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsMultiplexDocument,
-      variables: {
-        facilityId: "1",
-        result: "NEGATIVE",
-        pageNumber: 0,
-        pageSize: 20,
-      },
-    },
-    result: {
-      data: {
-        testResults: testResultsByResultValue,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetResultsCountByFacilityDocument,
-      variables: {
-        facilityId: "1",
-        role: "RESIDENT",
-      },
-    },
-    result: {
-      data: {
-        testResultsCount: testResultsByRole.length,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsMultiplexDocument,
-      variables: {
-        facilityId: "1",
-        role: "RESIDENT",
-        pageNumber: 0,
-        pageSize: 20,
-      },
-    },
-    result: {
-      data: {
-        testResults: testResultsByRole,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetResultsCountByFacilityDocument,
-      variables: {
-        facilityId: "1",
-        startDate: "2021-03-18T00:00:00.000Z",
-      },
-    },
-    result: {
-      data: {
-        testResultsCount: testResultsByStartDate.length,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsMultiplexDocument,
-      variables: {
-        facilityId: "1",
-        startDate: "2021-03-18T00:00:00.000Z",
-        pageNumber: 0,
-        pageSize: 20,
-      },
-    },
-    result: {
-      data: {
-        testResults: testResultsByStartDate,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetResultsCountByFacilityDocument,
-      variables: {
-        facilityId: "1",
-        startDate: "2021-03-18T00:00:00.000Z",
-        endDate: "2021-03-18T23:59:59.999Z",
-      },
-    },
-    result: {
-      data: {
-        testResultsCount: testResultsByStartDateAndEndDate.length,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsMultiplexDocument,
-      variables: {
-        facilityId: "1",
-        startDate: "2021-03-18T00:00:00.000Z",
-        endDate: "2021-03-18T23:59:59.999Z",
-        pageNumber: 0,
-        pageSize: 20,
-      },
-    },
-    result: {
-      data: {
-        testResults: testResultsByStartDateAndEndDate,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetResultsCountByFacilityDocument,
-      variables: {
-        facilityId: "2",
-      },
-    },
-    result: {
-      data: {
-        testResultsCount: testResultsByFacility.length,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsMultiplexDocument,
-      variables: {
-        facilityId: "2",
-        pageNumber: 0,
-        pageSize: 20,
-      },
-    },
-    result: {
-      data: {
-        testResults: testResultsByFacility,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetResultsCountByFacilityDocument,
-      variables: {
-        facilityId: null,
-      },
-    },
-    result: {
-      data: {
-        testResultsCount: testResultsByAllFacility.length,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsMultiplexDocument,
-      variables: {
-        facilityId: null,
-        pageNumber: 0,
-        pageSize: 20,
-      },
-    },
-    result: {
-      data: {
-        testResults: testResultsByAllFacility,
-      },
-    },
-  },
-  {
-    request: {
-      query: QUERY_PATIENT,
-      variables: {
-        facilityId: "1",
-        namePrefixMatch: "Cragell",
-        includeArchivedFacilities: true,
-      },
-    },
-    result: {
-      data: {
-        patients,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetResultsCountByFacilityDocument,
-      variables: {
-        facilityId: "1",
-      },
-    },
-    result: {
-      data: {
-        testResultsCount: testResults.length,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsMultiplexDocument,
-      variables: {
-        facilityId: "1",
-        pageNumber: 0,
-        pageSize: 20,
-      },
-    },
-    result: {
-      data: {
-        testResults,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetResultsCountByFacilityDocument,
-      variables: {
-        facilityId: "3",
-      },
-    },
-    result: {
-      data: {
-        testResultsCount: testResultsMultiplex.length,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetAllFacilitiesDocument,
-      variables: {
-        showArchived: true,
-      },
-    },
-    result: {
-      data: {
-        facilities: facilitiesIncludeArchived,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetFacilityResultsMultiplexDocument,
-      variables: {
-        facilityId: "3",
-        pageNumber: 0,
-        pageSize: 20,
-      },
-    },
-    result: {
-      data: {
-        testResults: testResultsMultiplex,
-      },
-    },
-  },
-  {
-    request: {
-      query: GetAllFacilitiesDocument,
-      variables: {
-        showArchived: false,
-      },
-    },
-    result: {
-      data: {
-        facilities,
-      },
-    },
-  },
-];
 
 describe("TestResultsList", () => {
   it("should render a list of tests", async () => {
