+++ frontend/src/app/testResults/mocks/patients.tsx
@@ -1,0 +1,13 @@
+export const patients = [
+  {
+    internalId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
+    firstName: "Barb",
+    middleName: "Whitaker",
+    lastName: "Cragell",
+    birthDate: "1960-11-07",
+    gender: "male",
+    role: "RESIDENT",
+    lookupId: null,
+    __typename: "Patient",
+  },
+];
