+++ frontend/src/app/testResults/mocks/resultsCSV.tsx
@@ -1,0 +1,164 @@
+const TEST_RESULTS_CSV = [
+  {
+    facility: { name: "Central Middle School", __typename: "Facility" },
+    dateTested: "2022-01-19T16:45:11.446Z",
+    result: "NEGATIVE",
+    correctionStatus: "ORIGINAL",
+    reasonForCorrection: null,
+    deviceType: {
+      name: "Cue",
+      manufacturer: "CueTM",
+      model: "Cue COVID Test",
+      swabType: "45345873",
+      __typename: "DeviceType",
+    },
+    patient: {
+      internalId: "48c523e8-7c65-4047-955c-e3f65bb8b58a",
+      firstName: "Barb",
+      middleName: "Whitaker",
+      lastName: "Cragell",
+      birthDate: "1960-11-07",
+      gender: "male",
+      role: "RESIDENT",
+      race: "asian",
+      ethnicity: "not_hispanic",
+      tribalAffiliation: [null],
+      lookupId: null,
+      telephone: "(905) 594-3725",
+      email: "fimevusuwu@mailinator.com",
+      street: "67 White Oak Avenue",
+      streetTwo: "Voluptatem optio in",
+      city: "Magna ut corrupti s",
+      county: "Maiores et sunt et q",
+      state: "NA",
+      zipCode: "00000",
+      country: null,
+      residentCongregateSetting: null,
+      employedInHealthcare: null,
+      preferredLanguage: null,
+      __typename: "Patient",
+    },
+    createdBy: {
+      nameInfo: {
+        firstName: "Nathan",
+        middleName: null,
+        lastName: "Carter",
+        __typename: "NameInfo",
+      },
+      __typename: "ApiUser",
+    },
+    symptoms:
+      '{"64531003":"false","103001002":"false","84229001":"false","68235000":"false","426000000":"false","49727002":"false","68962001":"false","422587007":"false","267036007":"false","62315008":"false","43724002":"false","36955009":"false","44169009":"false","422400008":"false","230145002":"false","25064002":"false","162397003":"false"}',
+    noSymptoms: false,
+    symptomOnset: null,
+    __typename: "TestResult",
+  },
+  {
+    facility: { name: "Central Middle School", __typename: "Facility" },
+    dateTested: "2022-01-19T16:42:46.744Z",
+    result: "NEGATIVE",
+    correctionStatus: "ORIGINAL",
+    reasonForCorrection: null,
+    deviceType: {
+      name: "Cue",
+      manufacturer: "CueTM",
+      model: "Cue COVID Test",
+      swabType: "45345873",
+      __typename: "DeviceType",
+    },
+    patient: {
+      firstName: "Cadman",
+      middleName: "Burton Conner",
+      lastName: "Day",
+      birthDate: "1903-04-07",
+      gender: "male",
+      race: "asian",
+      ethnicity: "not_hispanic",
+      tribalAffiliation: [null],
+      lookupId: null,
+      telephone: "(905) 594-3725",
+      email: "fimevusuwu@mailinator.com",
+      street: "67 White Oak Avenue",
+      streetTwo: "Voluptatem optio in",
+      city: "Magna ut corrupti s",
+      county: "Maiores et sunt et q",
+      state: "NA",
+      zipCode: "00000",
+      country: null,
+      role: "STUDENT",
+      residentCongregateSetting: null,
+      employedInHealthcare: null,
+      preferredLanguage: null,
+      __typename: "Patient",
+    },
+    createdBy: {
+      nameInfo: {
+        firstName: "Nathan",
+        middleName: null,
+        lastName: "Carter",
+        __typename: "NameInfo",
+      },
+      __typename: "ApiUser",
+    },
+    symptoms:
+      '{"64531003":"false","103001002":"false","84229001":"false","68235000":"false","426000000":"false","49727002":"false","68962001":"false","422587007":"false","267036007":"false","62315008":"false","43724002":"false","36955009":"false","44169009":"false","422400008":"false","230145002":"false","25064002":"false","162397003":"false"}',
+    noSymptoms: false,
+    symptomOnset: null,
+    __typename: "TestResult",
+  },
+  {
+    facility: { name: "Central Middle School", __typename: "Facility" },
+    dateTested: "2022-01-13T22:44:52.193Z",
+    result: "POSITIVE",
+    correctionStatus: "ORIGINAL",
+    reasonForCorrection: null,
+    deviceType: {
+      name: "Cue",
+      manufacturer: "CueTM",
+      model: "Cue COVID Test",
+      swabType: "45345873",
+      __typename: "DeviceType",
+    },
+    patient: {
+      firstName: "Bob",
+      middleName: null,
+      lastName: "International",
+      birthDate: "1970-09-09",
+      gender: "male",
+      race: null,
+      ethnicity: null,
+      tribalAffiliation: [null],
+      lookupId: null,
+      telephone: "(530) 867-5309",
+      email: "zedd@skylight.digital",
+      street: "123 Main St",
+      streetTwo: "",
+      city: null,
+      county: null,
+      state: "CA",
+      zipCode: "95820",
+      country: "USA",
+      role: "UNKNOWN",
+      residentCongregateSetting: null,
+      employedInHealthcare: null,
+      preferredLanguage: null,
+      __typename: "Patient",
+    },
+    createdBy: {
+      nameInfo: {
+        firstName: "Zedd",
+        middleName: null,
+        lastName: "Shmais",
+        __typename: "NameInfo",
+      },
+      __typename: "ApiUser",
+    },
+    symptoms:
+      '{"64531003":"false","103001002":"false","84229001":"false","68235000":"false","426000000":"false","49727002":"false","68962001":"false","422587007":"false","267036007":"false","62315008":"false","43724002":"false","36955009":"false","44169009":"false","422400008":"false","230145002":"false","25064002":"false","162397003":"false"}',
+    noSymptoms: false,
+    symptomOnset: null,
+    __typename: "TestResult",
+  },
+];
+
+export default TEST_RESULTS_CSV;
